requirements.txt
fastapi==0.115.9
python-dotenv==1.1.0
psycopg2==2.9.10
polars==1.26.0
pydantic==2.11.2
google-genai==1.9.0
uvicorn==0.34.0

